Elaborate on some properties of ABE
This is mostly free-written and probably leaves a lot of things
unaddressed and maybe doesn't say everything in the best way to
communicate where we really are and what we really know about ABE in
relation to traditional systems. But it's a start and hopefully will
prompt useful questions and criticisms.